Code split uncommon pages
