--- src/pages/taskHistory/index.ts
@@ -1,5 +1,0 @@
-import BuildVariantSelector from "./BuildVariantSelector";
-import ColumnHeaders from "./ColumnHeaders";
-import TaskHistoryRow from "./TaskHistoryRow";
-
-export { BuildVariantSelector, TaskHistoryRow, ColumnHeaders };
